% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_accuracy}
\alias{plot_accuracy}
\title{If we target N people, what fraction of the true poor would receive funds?
True Positives / Total Positives
Note that this is a reparameterization of the ROC curve}
\usage{
plot_accuracy(THRESHOLD, ..., DISPLAY_TRUE = FALSE, DISPLAY_CUTOFFS = FALSE,
  POINT_COUNT = 20)
}
\description{
If we target N people, what fraction of the true poor would receive funds?
True Positives / Total Positives
Note that this is a reparameterization of the ROC curve
}

