% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_accuracy_dollars}
\alias{plot_accuracy_dollars}
\title{With a fixed amount of money, if we target N people, what fraction would go to the true poor?
True Positives / (True Positives + False Positives)}
\usage{
plot_accuracy_dollars(THRESHOLD, ..., DISPLAY_TRUE = FALSE,
  DISPLAY_CUTOFFS = FALSE, POINT_COUNT = 20)
}
\description{
With a fixed amount of money, if we target N people, what fraction would go to the true poor?
True Positives / (True Positives + False Positives)
}

