% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_roc}
\alias{plot_roc}
\title{Produce an ROC curve which plots a given method's sensitivity/specificity with respect
a given poverty threshold.}
\usage{
plot_roc(THRESHOLD, ..., PLOT_FOLDS = FALSE)
}
\description{
Produce an ROC curve which plots a given method's sensitivity/specificity with respect
a given poverty threshold.
}

