% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{standardize_predictors}
\alias{standardize_predictors}
\title{#############
### This file contains ML prediction functions for different algorithms,
### designed to be harmonized
#############}
\usage{
standardize_predictors(df, target)
}
\description{
#############
### This file contains ML prediction functions for different algorithms,
### designed to be harmonized
#############
}

