% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model.R
\name{standardize_x}
\alias{standardize_x}
\title{Standardize a dataset so that predictor variables have unit variance.}
\usage{
standardize_x(yx)
}
\arguments{
\item{yx}{A dataframe. The first column should be the variable to be predicted.}
}
\description{
Standardize a dataset so that predictor variables have unit variance.
}

